REFACTOR util function을 banhyang.core.utils.py로 이동
